refactor(memory): gather leaf v1 modules under v1/
