+++ assistant/src/__tests__/injector-pkb-v2-silenced.test.ts
@@ -48,7 +48,7 @@ mock.module("../plugins/defaults/memory/config.js", () => ({
   }),
 }));
 
-mock.module("../plugins/defaults/memory/pkb/pkb-search.js", () => ({
+mock.module("../plugins/defaults/memory/v1/pkb/pkb-search.js", () => ({
   searchPkbFiles: async () => [],
 }));
 
@@ -56,7 +56,7 @@ const { applyRuntimeInjections } =
   await import("../daemon/conversation-runtime-assembly.js");
 const { registerDefaultPluginInjectors } =
   await import("../plugins/defaults/index.js");
-import { getPkbRoot } from "../plugins/defaults/memory/pkb/types.js";
+import { getPkbRoot } from "../plugins/defaults/memory/v1/pkb/types.js";
 import type { TurnContext } from "../plugins/types.js";
 import type { Message } from "../providers/types.js";
 import { getWorkspacePromptPath } from "../util/platform.js";
