+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -87,6 +87,11 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../../persistence/db-connection.js",
     "../../../../../persistence/embeddings/embed.js",
     "../../../../../persistence/embeddings/embedding-backend.js",
+    "../../../../../persistence/embeddings/embedding-types.js",
+    "../../../../../persistence/embeddings/qdrant-circuit-breaker.js",
+    "../../../../../persistence/embeddings/qdrant-client.js",
+    "../../../../../persistence/job-utils.js",
+    "../../../../../persistence/jobs-store.js",
     "../../../../../persistence/raw-query.js",
     "../../../../../persistence/schema/index.js",
     "../../../../../security/untrusted-content.js",
@@ -127,7 +132,9 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../persistence/raw-query.js",
     "../../../../persistence/schema/conversations.js",
     "../../../../persistence/schema/index.js",
+    "../../../../prompts/persona-resolver.js",
     "../../../../prompts/system-prompt.js",
+    "../../../../prompts/template-detection.js",
     "../../../../providers/platform-proxy/context.js",
     "../../../../runtime/assistant-event-hub.js",
     "../../../../runtime/auth/route-policy.js",
@@ -188,11 +195,9 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../persistence/schema/index.js",
     "../../../prompts/persona-resolver.js",
     "../../../prompts/system-prompt.js",
-    "../../../prompts/template-detection.js",
     "../../../runtime/actor-trust-resolver.js",
     "../../../runtime/agent-wake.js",
     "../../../runtime/assistant-stream-state.js",
-    "../../../runtime/background-job-runner.js",
     "../../../runtime/capabilities.js",
     "../../../runtime/services/auto-analysis-guard.js",
     "../../../runtime/sync/resource-sync-events.js",
