+++ assistant/src/__tests__/rebuild-index-graph-nodes.test.ts
@@ -124,7 +124,7 @@ mock.module("../persistence/db-connection.js", () => ({
 
 // ── Tests ────────────────────────────────────────────────────────────
 
-import { rebuildIndexJob } from "../plugins/defaults/memory/job-handlers/index-maintenance.js";
+import { rebuildIndexJob } from "../plugins/defaults/memory/v1/job-handlers/index-maintenance.js";
 
 describe("rebuildIndexJob", () => {
   beforeEach(() => {
