+++ assistant/src/plugins/defaults/memory/v1/filing-jobs.ts
@@ -17,9 +17,9 @@
 
 import type { LLMCallSite } from "@vellumai/plugin-api";
 
-import type { MemoryJob } from "../../../persistence/jobs-store.js";
-import { runBackgroundJob } from "../../../runtime/background-job-runner.js";
-import { getLogger } from "./logging.js";
+import type { MemoryJob } from "../../../../persistence/jobs-store.js";
+import { runBackgroundJob } from "../../../../runtime/background-job-runner.js";
+import { getLogger } from "../logging.js";
 import { hasPkbBufferContent } from "./pkb-schedule.js";
 
 const log = getLogger("filing-jobs");
