+++ assistant/src/plugins/defaults/memory/v1/identity-context.ts
@@ -1,7 +1,7 @@
-import { resolveGuardianPersona } from "../../../prompts/persona-resolver.js";
-import { readPromptFile } from "../../../prompts/system-prompt.js";
-import { isTemplateContent } from "../../../prompts/template-detection.js";
-import { getWorkspacePromptPath } from "./paths.js";
+import { resolveGuardianPersona } from "../../../../prompts/persona-resolver.js";
+import { readPromptFile } from "../../../../prompts/system-prompt.js";
+import { isTemplateContent } from "../../../../prompts/template-detection.js";
+import { getWorkspacePromptPath } from "../paths.js";
 
 const IDENTITY_PROMPT_FILES = ["IDENTITY.md", "SOUL.md"] as const;
