+++ assistant/src/plugins/defaults/memory/v1/job-handlers/embedding.test.ts
@@ -13,7 +13,7 @@ const embedAndUpsertCalls: Array<{
   extraPayload: unknown;
 }> = [];
 
-mock.module("../../../../persistence/job-utils.js", () => ({
+mock.module("../../../../../persistence/job-utils.js", () => ({
   asString: (value: unknown) =>
     typeof value === "string" && value.length > 0 ? value : null,
   embedAndUpsert: async (
@@ -33,11 +33,11 @@ mock.module("../../../../persistence/job-utils.js", () => ({
   },
 }));
 
-import { resetDbForTesting } from "../../../../__tests__/db-test-helpers.js";
-import { getDb } from "../../../../persistence/db-connection.js";
-import { initializeDb } from "../../../../persistence/db-init.js";
-import type { MemoryJob } from "../../../../persistence/jobs-store.js";
-import { mediaAssets } from "../../../../persistence/schema/index.js";
+import { resetDbForTesting } from "../../../../../__tests__/db-test-helpers.js";
+import { getDb } from "../../../../../persistence/db-connection.js";
+import { initializeDb } from "../../../../../persistence/db-init.js";
+import type { MemoryJob } from "../../../../../persistence/jobs-store.js";
+import { mediaAssets } from "../../../../../persistence/schema/index.js";
 import { embedMediaJob } from "./embedding.js";
 
 describe("embedMediaJob", () => {
