+++ assistant/src/plugins/defaults/memory/v1/job-handlers/embedding.ts
@@ -3,17 +3,17 @@ import { readFile } from "node:fs/promises";
 import { embedAndUpsert } from "@vellumai/plugin-api";
 import { and, eq } from "drizzle-orm";
 
-import { getDb } from "../../../../persistence/db-connection.js";
-import type { EmbeddingInput } from "../../../../persistence/embeddings/embedding-types.js";
-import { asString } from "../../../../persistence/job-utils.js";
-import type { MemoryJob } from "../../../../persistence/jobs-store.js";
+import { getDb } from "../../../../../persistence/db-connection.js";
+import type { EmbeddingInput } from "../../../../../persistence/embeddings/embedding-types.js";
+import { asString } from "../../../../../persistence/job-utils.js";
+import type { MemoryJob } from "../../../../../persistence/jobs-store.js";
 import {
   mediaAssets,
   memorySegments,
   memorySummaries,
   messages,
-} from "../../../../persistence/schema/index.js";
-import { extractMediaBlocks } from "../message-media.js";
+} from "../../../../../persistence/schema/index.js";
+import { extractMediaBlocks } from "../../message-media.js";
 
 export async function embedSegmentJob(job: MemoryJob): Promise<void> {
   const segmentId = asString(job.payload.segmentId);
