+++ assistant/src/plugins/defaults/memory/v1/pkb-schedule.ts
@@ -10,8 +10,8 @@
 import { existsSync, readFileSync } from "node:fs";
 import { join } from "node:path";
 
-import { stripCommentLines } from "./host-utils.js";
-import { getWorkspaceDir } from "./paths.js";
+import { stripCommentLines } from "../host-utils.js";
+import { getWorkspaceDir } from "../paths.js";
 
 /**
  * Whether `pkb/buffer.md` has any filable content (comment lines and
