+++ assistant/src/plugins/defaults/memory/v1/pkb/context.ts
@@ -1,8 +1,8 @@
 import { existsSync, readFileSync } from "node:fs";
 import { join, resolve } from "node:path";
 
-import { stripCommentLines } from "../host-utils.js";
-import { getWorkspaceDir } from "../paths.js";
+import { stripCommentLines } from "../../host-utils.js";
+import { getWorkspaceDir } from "../../paths.js";
 import { getPkbAutoInjectList } from "./autoinject.js";
 
 /** Max buffer.md lines injected into prompts — keeps context bounded even when filing is off. */
