+++ assistant/src/plugins/defaults/memory/v1/pkb/pkb-index.ts
@@ -17,8 +17,8 @@ import { join, relative } from "node:path";
 
 import { embedAndUpsert } from "@vellumai/plugin-api";
 
-import { withQdrantBreaker } from "../../../../persistence/embeddings/qdrant-circuit-breaker.js";
-import { getQdrantClient } from "../../../../persistence/embeddings/qdrant-client.js";
+import { withQdrantBreaker } from "../../../../../persistence/embeddings/qdrant-circuit-breaker.js";
+import { getQdrantClient } from "../../../../../persistence/embeddings/qdrant-client.js";
 import type { PkbIndexEntry } from "./types.js";
 import { PKB_TARGET_TYPE } from "./types.js";
