+++ assistant/src/plugins/defaults/memory/v1/pkb/pkb-reconcile.ts
@@ -13,9 +13,9 @@
 
 import { join } from "node:path";
 
-import { getQdrantClient } from "../../../../persistence/embeddings/qdrant-client.js";
+import { getQdrantClient } from "../../../../../persistence/embeddings/qdrant-client.js";
+import { getLogger } from "../../logging.js";
 import { enqueuePkbIndexJob } from "../jobs/embed-pkb-file.js";
-import { getLogger } from "../logging.js";
 import { deletePkbFilePoints, scanPkbFiles } from "./pkb-index.js";
 import { PKB_TARGET_TYPE } from "./types.js";
